Moving modules out of subdir
moved modules out of nested subdir for simpler structure
